fix(celt): keep SIMD benchmark seeds 32-bit portable
int32(i*2654435761 + 1) does not compile on 32-bit platforms (GOARCH=386,
arm): the untyped constant 2654435761 overflows a 32-bit int. Multiply in
uint32 instead (matching pitch_simd_test.go's existing idiom) so the seed
wraps in uint32 and stays portable. Fixes the new denormaliseGain benchmark
and the sibling haar1 benchmark it was copied from.

Verified with GOARCH=386 and GOARCH=arm go vet.

+++ internal/celt/denormalise_simd_test.go
@@ -182,7 +182,7 @@ func BenchmarkDenormaliseGain(b *testing.B) {
 	for _, n := range lengths {
 		seed := make([]int32, n)
 		for i := range seed {
-			seed[i] = int32(i*2654435761 + 1)
+			seed[i] = int32(uint32(i)*2654435761 + 1)
 		}
 		const g = int32(1 << 24)
 		b.Run(fmt.Sprintf("scalar/n%d", n), func(b *testing.B) {
